# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zytrixon Mobile API Documentation",
        "description": "API documentation for the Zytrixon Truck Billing mobile application.",
        "contact": {
            "email": "admin@zytrixon.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/drivers": {
            "get": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Get all drivers",
                "operationId": "b7e92c879078ab66ae9f678f5c362697",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Drivers retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Create a new driver",
                "operationId": "af273841718d9047cc3bfb8800ddf839",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "mobile",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Driver John"
                                    },
                                    "mobile": {
                                        "type": "string",
                                        "example": "9876543210"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "driver@zytrixon.com",
                                        "nullable": true
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "license_number": {
                                        "type": "string",
                                        "example": "DL-123456789",
                                        "nullable": true
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Delhi",
                                        "nullable": true
                                    },
                                    "license_document": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Driver created successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/expense-categories": {
            "get": {
                "tags": [
                    "Expense Categories"
                ],
                "summary": "Get all expense categories",
                "operationId": "7775b3b56b348d9fe63d545529fcb5c4",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categories retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Expense Categories"
                ],
                "summary": "Create a new expense category",
                "operationId": "69dcd1e98d786384d56af5bbd9ec374b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Fuel"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/expense-categories/{id}": {
            "delete": {
                "tags": [
                    "Expense Categories"
                ],
                "summary": "Delete an expense category",
                "operationId": "50c0c439731d3f9e64cc64e8e0bcab9d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum token here",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}